test_obfuscator/requirements.txt
torch>=1.10.0
transformers>=4.20.0
tqdm>=4.62.0
requests>=2.27.1
pandas>=1.3.5
matplotlib>=3.5.0
seaborn>=0.11.2
python-Levenshtein>=0.12.0
pyyaml>=6.0
uvicorn>=0.17.0
numpy>=1.21.0
fastapi>=0.70.0
scikit-learn>=1.0.0 